requirements.txt
amqp==5.2.0
apispec==6.4.0
async-timeout==4.0.3
billiard==4.2.0
blinker==1.7.0
celery==5.3.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
datadispatch==1.0.0
Flask==3.0.1
flask-smorest==0.42.3
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
kombu==5.3.5
MarkupSafe==2.1.4
marshmallow==3.20.2
packaging==23.2
prompt-toolkit==3.0.43
python-dateutil==2.8.2
python-dotenv==1.0.1
redis==5.0.1
requests==2.31.0
six==1.16.0
tasks==2.8.0
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.0
vine==5.1.0
wcwidth==0.2.13
webargs==8.4.0
Werkzeug==3.0.1
zipp==3.17.0
pytz==2022.5
mysql-connector-python==8.0.28
unidecode==1.3.2
